# Conflicts:
#	identityprovider/img/profileImage
#	portfolio/src/main/resources/templates/fragments/eventModal.html

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/ViewUsersController.java
@@ -1,18 +1,16 @@
 package nz.ac.canterbury.seng302.portfolio.controller;
 
 import nz.ac.canterbury.seng302.portfolio.model.UserSorting;
-import nz.ac.canterbury.seng302.portfolio.service.ElementService;;
-import nz.ac.canterbury.seng302.portfolio.service.RegisterClientService;
-import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
-import nz.ac.canterbury.seng302.portfolio.service.UserSortingService;
+import nz.ac.canterbury.seng302.portfolio.service.*;;
 import nz.ac.canterbury.seng302.shared.identityprovider.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import javax.management.relation.Role;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Objects;
 
@@ -24,6 +22,9 @@ public class ViewUsersController {
 
     @Autowired
     private RegisterClientService registerClientService;
+
+    @Autowired
+    private PermissionService permissionService;
 
     @Autowired
     private UserAccountClientService userAccountClientService;
@@ -45,26 +46,21 @@ public class ViewUsersController {
     @GetMapping("/viewUsers")
     public String showUserTablePage(
             Model model,
+            HttpServletRequest request,
             @AuthenticationPrincipal AuthState principal
     ) {
         UserResponse getUserByIdReply;
         Integer id = userAccountClientService.getUserIDFromAuthState(principal);
         elementService.addHeaderAttributes(model, id);
         getUserByIdReply = registerClientService.getUserData(id);
-        String role = principal.getClaimsList().stream()
-                .filter(claim -> claim.getType().equals("role"))
-                .findFirst()
-                .map(ClaimDTO::getValue)
-                .orElse("NOT FOUND");
-
-
+        String role = elementService.getUserHighestRole(getUserByIdReply);
+        elementService.addDeniedMessage(model, request);
         model.addAttribute("currentUserRole", role);
         model.addAttribute("currentUsername", getUserByIdReply.getUsername());
         model.addAttribute("userId", id);
         model.addAttribute("studentRole", UserRole.STUDENT);
         model.addAttribute("teacherRole", UserRole.TEACHER);
         model.addAttribute("adminRole", UserRole.COURSE_ADMINISTRATOR);
-
         PaginatedUsersResponse response = userAccountClientService.getAllUsers();
         userResponseList = response.getUsersList();
         model.addAttribute("users", userResponseList);
@@ -86,9 +82,9 @@ public class ViewUsersController {
      * @return viewUsers html page
      */
     @RequestMapping(value="/viewUsers/saveSort", method=RequestMethod.POST)
-    public String updateSprintRangeErrors(@RequestParam(value="columnIndex") Integer columnIndex,
-                                          @RequestParam(value="sortOrder") String sortOrder,
-                                          @AuthenticationPrincipal AuthState principal) {
+    public String saveSort(@RequestParam(value="columnIndex") Integer columnIndex,
+                           @RequestParam(value="sortOrder") String sortOrder,
+                           @AuthenticationPrincipal AuthState principal) {
         Integer id = userAccountClientService.getUserIDFromAuthState(principal);
         UserSorting userSorting = new UserSorting(id, columnIndex, sortOrder);
         userSortingService.updateUserSorting(userSorting);
@@ -102,20 +98,31 @@ public class ViewUsersController {
      * @return list of users page(html)
      */
     @RequestMapping(value="/add_role", method=RequestMethod.POST)
-    public String addRole(
+    public String addRole(Model model,
+                              RedirectAttributes rm,
                               @RequestParam(value = "role") String role,
-                              @RequestParam(value = "userId") int userId
-    ) {
+                              @RequestParam(value = "userId") int userId,
+                              @AuthenticationPrincipal AuthState principal
+                              ) {
 
-        if (role.equals("student")) {
-            UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.STUDENT);
-        } else if (role.equals("teacher")) {
-            UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.TEACHER);
-        } else {
-            UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.COURSE_ADMINISTRATOR);
+        Integer id = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, id);
+
+        // Check if current user's operation is valid, if invalid, access denied error is displayed to user
+        if (permissionService.isValidToModifyRole(role, id)) {
+            if (role.equals("student")) {
+                UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.STUDENT);
+            } else if (role.equals("teacher")) {
+                UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.TEACHER);
+            } else {
+                UserRoleChangeResponse roleChangeResponse = userAccountClientService.addRoleToUser(userId, UserRole.COURSE_ADMINISTRATOR);
+            }
+            return "redirect:viewUsers";
         }
+        rm.addFlashAttribute("isAccessDenied", true);
         return "redirect:viewUsers";
     }
+
 
     /***
      * POST method request handler when the url is "/delete_role"
@@ -127,22 +134,33 @@ public class ViewUsersController {
     @RequestMapping(value="/delete_role", method=RequestMethod.POST)
     public String deleteRole(Model model,
                               @RequestParam(value = "deletedRole") String role,
-                              @RequestParam(value = "userId") int userId
-    ) {
-        UserRoleChangeResponse roleChangeResponse;
-        if (Objects.equals(role, "STUDENT")) {
-            roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.STUDENT);
-        } else if (Objects.equals(role, "TEACHER")) {
-            roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.TEACHER);
-        } else {
-            roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.COURSE_ADMINISTRATOR);
-        }
-        if (roleChangeResponse.getIsSuccess()) {
-            return "redirect:viewUsers";
-        } else {
-            model.addAttribute("errorMessage", "Error deleting user");
-            return "redirect:error";
-        }
+                              @RequestParam(value = "userId") int userId,
+                             @AuthenticationPrincipal AuthState principal
+                             ) {
 
+        Integer id = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, id);
+
+        // Check if current user's operation is valid, if invalid, access denied error is displayed to user
+        if (permissionService.isValidToModifyRole(role, id)) {
+            UserRoleChangeResponse roleChangeResponse;
+            if (Objects.equals(role, "STUDENT")) {
+                roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.STUDENT);
+            } else if (Objects.equals(role, "TEACHER")) {
+                roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.TEACHER);
+            } else {
+                roleChangeResponse = userAccountClientService.deleteRoleFromUser(userId, UserRole.COURSE_ADMINISTRATOR);
+            }
+            if (roleChangeResponse.getIsSuccess()) {
+                return "redirect:viewUsers";
+            } else {
+                model.addAttribute("errorMessage", "Error deleting user");
+                return "redirect:error";
+            }
+        }
+//        rm.addFlashAttribute("isAccessDenied", true);
+        return "redirect:viewUsers";
     }
+
+
 }
